refactor post edit
