+++ var/Widget/Contents/Attachment/Edit.php
@@ -7,6 +7,7 @@ use Typecho\Widget\Exception;
 use Typecho\Widget\Helper\Form;
 use Typecho\Widget\Helper\Layout;
 use Widget\ActionInterface;
+use Widget\Contents\EditTrait;
 use Widget\Contents\Post\Edit as PostEdit;
 use Widget\Notice;
 use Widget\Upload;
@@ -26,6 +27,8 @@ if (!defined('__TYPECHO_ROOT_DIR__')) {
  */
 class Edit extends PostEdit implements ActionInterface
 {
+    use EditTrait;
+
     /**
      * 执行函数
      *
@@ -35,20 +38,6 @@ class Edit extends PostEdit implements ActionInterface
     {
         /** 必须为贡献者以上权限 */
         $this->user->pass('contributor');
-
-        /** 获取文章内容 */
-        if (!empty($this->request->cid)) {
-            $this->db->fetchRow($this->select()
-                ->where('table.contents.type = ?', 'attachment')
-                ->where('table.contents.cid = ?', $this->request->filter('int')->get('cid'))
-                ->limit(1), [$this, 'push']);
-
-            if (!$this->have()) {
-                throw new Exception(_t('文件不存在'), 404);
-            } elseif (!$this->allow('edit')) {
-                throw new Exception(_t('没有编辑权限'), 403);
-            }
-        }
     }
 
     /**
@@ -229,32 +218,7 @@ class Edit extends PostEdit implements ActionInterface
         $posts = $this->request->filter('int')->getArray('cid');
         $deleteCount = 0;
 
-        foreach ($posts as $post) {
-            // 删除插件接口
-            self::pluginHandle()->call('delete', $post, $this);
-
-            $condition = $this->db->sql()->where('cid = ?', $post);
-            $row = $this->db->fetchRow($this->select()
-                ->where('table.contents.type = ?', 'attachment')
-                ->where('table.contents.cid = ?', $post)
-                ->limit(1), [$this, 'push']);
-
-            if ($this->isWriteable(clone $condition) && $this->delete($condition)) {
-                /** 删除文件 */
-                Upload::deleteHandle($row);
-
-                /** 删除评论 */
-                $this->db->query($this->db->delete('table.comments')
-                    ->where('cid = ?', $post));
-
-                // 完成删除插件接口
-                self::pluginHandle()->call('finishDelete', $post, $this);
-
-                $deleteCount++;
-            }
-
-            unset($condition);
-        }
+        $this->deleteByIds($posts, $deleteCount);
 
         if ($this->request->isAjax()) {
             $this->response->throwJson($deleteCount > 0 ? ['code' => 200, 'message' => _t('文件已经被删除')]
@@ -291,32 +255,7 @@ class Edit extends PostEdit implements ActionInterface
                 ->page($page, 100)), 'cid');
             $page++;
 
-            foreach ($posts as $post) {
-                // 删除插件接口
-                self::pluginHandle()->call('delete', $post, $this);
-
-                $condition = $this->db->sql()->where('cid = ?', $post);
-                $row = $this->db->fetchRow($this->select()
-                    ->where('table.contents.type = ?', 'attachment')
-                    ->where('table.contents.cid = ?', $post)
-                    ->limit(1), [$this, 'push']);
-
-                if ($this->isWriteable(clone $condition) && $this->delete($condition)) {
-                    /** 删除文件 */
-                    Upload::deleteHandle($row);
-
-                    /** 删除评论 */
-                    $this->db->query($this->db->delete('table.comments')
-                        ->where('cid = ?', $post));
-
-                    // 完成删除插件接口
-                    self::pluginHandle()->call('finishDelete', $post, $this);
-
-                    $deleteCount++;
-                }
-
-                unset($condition);
-            }
+            $this->deleteByIds($posts, $deleteCount);
         } while (count($posts) == 100);
 
         /** 设置提示信息 */
@@ -339,8 +278,45 @@ class Edit extends PostEdit implements ActionInterface
     {
         $this->security->protect();
         $this->on($this->request->is('do=delete'))->deleteAttachment();
-        $this->on($this->have() && $this->request->is('do=update'))->updateAttachment();
+        $this->on($this->request->is('do=update'))
+            ->prepareEdit('attachment', false, _t('文件不存在'))
+            ->updateAttachment();
         $this->on($this->request->is('do=clear'))->clearAttachment();
         $this->response->redirect($this->options->adminUrl);
+    }
+
+    /**
+     * @param array $posts
+     * @param int $deleteCount
+     * @return void
+     */
+    protected function deleteByIds(array $posts, int &$deleteCount): void
+    {
+        foreach ($posts as $post) {
+            // 删除插件接口
+            self::pluginHandle()->call('delete', $post, $this);
+
+            $condition = $this->db->sql()->where('cid = ?', $post);
+            $row = $this->db->fetchRow($this->select()
+                ->where('table.contents.type = ?', 'attachment')
+                ->where('table.contents.cid = ?', $post)
+                ->limit(1), [$this, 'push']);
+
+            if ($this->isWriteable(clone $condition) && $this->delete($condition)) {
+                /** 删除文件 */
+                Upload::deleteHandle($row);
+
+                /** 删除评论 */
+                $this->db->query($this->db->delete('table.comments')
+                    ->where('cid = ?', $post));
+
+                // 完成删除插件接口
+                self::pluginHandle()->call('finishDelete', $post, $this);
+
+                $deleteCount++;
+            }
+
+            unset($condition);
+        }
     }
 }
